refactor: import문 최적화 및 메서드 이름 변경
