+++ src/test/java/leaguehub/leaguehubbackend/controller/ParticipantControllerTest.java
@@ -11,6 +11,7 @@ import leaguehub.leaguehubbackend.domain.channel.entity.ChannelRule;
 import leaguehub.leaguehubbackend.domain.member.entity.Member;
 import leaguehub.leaguehubbackend.domain.participant.entity.Participant;
 import leaguehub.leaguehubbackend.domain.participant.service.ParticipantManagementService;
+import leaguehub.leaguehubbackend.domain.participant.service.ParticipantRoleAndPermissionService;
 import leaguehub.leaguehubbackend.fixture.ChannelFixture;
 import leaguehub.leaguehubbackend.fixture.ParticipantFixture;
 import leaguehub.leaguehubbackend.fixture.UserFixture;
@@ -70,6 +71,8 @@ class ParticipantControllerTest {
 
     @Autowired
     ParticipantManagementService participantManagementService;
+    @Autowired
+    ParticipantRoleAndPermissionService participantRoleAndPermissionService;
 
     @Autowired
     ObjectMapper mapper;
@@ -538,7 +541,7 @@ class ParticipantControllerTest {
             dummyParticipant.approveParticipantMatch();
         }
         Participant dummy = getParticipant("DummyName1", channel, "DummyGameId1", "DummyNickname1");
-        participantService.approveParticipantRequest(channel.getChannelLink(), dummy.getId());
+        participantRoleAndPermissionService.approveParticipantRequest(channel.getChannelLink(), dummy.getId());
 
         Participant dummy1 = getParticipant("DummyName2", channel, "DummyGameId2", "DummyNickname2");
